requirements.txt
requests
selenium
pypdf
python-docx
chromadb
sentence-transformers
